refactor : 브랜드별 상품 조회 별도의 컨트롤러로 분리
